requirements.txt
streamlit
numpy
pandas
opencv-python
tensorflow==2.15.0
pillow